move `_get_binary_filename` to pippy.utils to prevent circular imports
