+++ pippy/hf/_SaveModule.py
@@ -10,6 +10,7 @@ import torch
 import torch.distributed as dist
 
 from pippy.IR import Pipe
+from pippy.utils import _get_binary_filename
 
 CKPT_INDEX_JSON_FILENAME = "pytorch_model.bin.index.json"
 
@@ -108,26 +109,6 @@ def _save_index(
     _atomic_write(json_str, filepath)
 
     logging.info(f"Saved index file to {filepath}")
-
-
-def _get_binary_filename(cur_idx: int, is_optim: bool = False) -> str:  # type: ignore[valid-type]
-    """
-    Gets filename for pytorch checkpoint binary based on current index and world size.
-
-    Args:
-        cur_idx (int): current device index
-        is_optim (bool): True if generating binary filename for optimizer,
-                         False otherwise
-
-    Returns:
-        str: checkpoint filename
-    """
-    idx = str(cur_idx + 1).zfill(5)
-    world_size = str(dist.get_world_size()).zfill(5)
-
-    state_type = "optim" if is_optim else "model"
-
-    return f"pytorch_{state_type}-{idx}-of-{world_size}.bin"
 
 
 def _save_params(submod: torch.nn.Module, checkpoint_dir: str) -> None:
